I am releasing the restriction that uninit_encounter_func can't set CommData.ConversationPhrases to null.
Now, it's impossible at this time for CommData to be set to null, since &CommData is never passed to any functions, and it was copied by value from the LOCDATA* returned by the various comm init functions.

Still, this doesn't hurt and if we change things around in an appropriate way - adding &CommData to uninit_encounter_func's argument list, say - it will help.